[SYCL][E2E] Add option to build tests on run-only mode if marked as `REQUIRES: build-and-run-mode` (#16306)
Prior if a test was marked as `REQUIRES: build-and-run-mode` it would be
reported as unsupported in `run-only` mode, because there would be no
test binaries created from `build-only` mode to be able to run the test.

When calling lit with `--param test-mode=run-only --param
fallback-to-build-if-requires-build-and-run=True`, tests that are marked
with `REQUIRES: build-and-run-mode` will override the `RUN:` line
filtering, and run all lines, including those that compile the test
binaries. This makes it so we still have coverage of tests marked as
exceptions for splitting on the `run-only` system.