[SYCL] Keep the old format of files in persistent cache
PR#16056 fixed multi-device support for persistent cache but also
changed format of the files in persistent cache - we used to write
number of binaries before binary size and binary data but stopped doing that
because we always have a single binary (for a single device) per file. Because of that
new runtime stopped working with persistent cache created by older runtimes.

This PR restores the old format of files in persistent cache.

For that we need to write number of binaries to the file (before binary size and binary data)
though it is always equal to 1 in current implementation. Even in the old implementation
we could only put a single binary to the persistent cache in all scenarios, multi-device
case wasn't supported, didn't have an API to create a program from multiple binaries.
So we can assert that number of binaries is always 1 when reading from the cache.