[SYCL][Doc] Initial spec conversion to Sphinx
The upstream clang project uses reStructuredText and Sphinx for its
documentation, so it makes sense to use these doc tools also for the
SYCL extension specifications.  Converting these documents to Sphinx
should remove one of the barriers to upstreaming our SYCL extensions.

This PR converts only a few specifications in order to get feedback on
the style and document structure.

I also took this as an opportunity to update the template we use for
writing new specifications.  We've gained a lot of experience since the
template was first written, and these updates demonstrate our new
preferred style for writing specifications.  I've also expanded the
template to show sample formatting for common scenarios that occur when
writing SYCL extensions.  For example, there is a sample section showing
an extension that adds member functions to an existing class, a sample
section showing an extension that adds new non-member functions, etc.