[CI][SYCL][E2E] Split gen12 pre-commit testing into build and run stages (#16321)
This patch enables E2E build and run splitting on pre-commit CI.
Currently one of the main bottlenecks in pre-commit CI is the Linux
Gen12 testing which takes between 30 ~ 40 minutes to complete. By
instead building these tests on our fast build systems, and then
transferring the binaries to the Gen12 machine we can cut down on this
time by roughly half.

Sadly, the Windows Gen12 testing will still be the main bottleneck so
total time for pre-commit won't improve with this PR, but we will at
least get faster feedback for gen12 on Linux.