CUTensorMap is only in CUDA v12
The documentation at https://docs.nvidia.com/cuda/cutensor/latest/ say
it's supported in 11.0, 11.8, and 12.x

Our local testing against 11.7 fails because none of the required types
or functions are declared in the header.

It's safer to remain buildable with older versions but with this feature
disabled than requiring a version bump that isn't otherwise warranted.