[CI] Run E2E tests on PVC on postcommit (#15308)
This PR enables running E2E tests on PVC in the post-commit.

Currently, the latest PVC driver hangs when a large number of jobs are
submitted to it in parallel. Due to this, E2E tests might timeout
spuriously on PVC. The PVC bug is being tracked internally via
GSD-10241. To prevent spurious test timeouts, we restrict the number of
parallel E2E test runs on PVC and to prevent disruption in CI, we are
currently enabling PVC in post-commit only. When the bug gets fixed, we
can move PVC testing to pre-commit.