[SYCL] Fix use of memcpy in group_load
The implementation of group_load currently uses std::memcpy, which is
not guaranteed to be supported in kernel code. This can, for example,
manifest if the user sets -D_FORTIFY_SOURCE=2 as it may insert an
assertion into the code, which in turn can fail to compile for devices.
This commit fixes this by using sycl::detail::memcpy_no_adl instead.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>