[NFC][SYCL][Driver] Add missing msan device libraries for testing (#16446)
For driver testing, we rely on the existance of device libraries to
satisfy the pulling in of required device libraries. When the compiler
is built without the libraries, testing needs to use the specific test
inputs to resolve.

Add the missing msan device libraries to fix failing tests when the
device libraries are not built (e.g. sanitizer specific builds)