[SYCL][E2E] Add XFAIL trackers for improperly XFAIL-ed E2E tests (#16421)
Add XFAIL trackers for improperly XFAIL-ed E2E tests. This brings the
number of improperly XFAIL-ed tests down to zero.
For the tests below, the XFAILs have been removed entirely.

- For `sycl/test-e2e/ESIMD/hardware_dispatch.cpp`, the XFAIL is on
igc-dev but the test requires GEN12 which does not have igc-dev so this
XFAIL is meaningless. Thanks to @sarnex for noting this.

- Three tests under `test-e2e/Reduction` which are XFAIL-ed only on
hip_nvidia which we do not support anymore.

Closes: https://github.com/intel/llvm/issues/16400