[Driver][SYCL] warning emitted when compiling wrapped object (#16397)
After the switch to using clang to build the wrapped object the embedded
target in the IR is caused a warning to be emitted. Fix this by
adjusting the clang-offload-wrapper call to use the same host triple
that is used when performing a regular compilation. Matching these up
allows for no diagnostic to be emitted.

Warning emitted: warning: overriding the module target triple with
      x86_64-pc-windows-msvc19.29.30154 [-Woverride-module]