[SYCL] ProgramManager: Remove invalidated entries in NativePrograms before insert (#15973)
Fixes https://github.com/intel/llvm/issues/14972

This commit erases existing entries with ur handle that is just
created/returned by backend. All existing entries in NativePrograms are
known to be invalid in this case.
Could not erase them on UrProgramRelease call since we have no tracking
of program handle references on SYCL RT level and it is not feasible to
add it. Obtaining ref count from ur is not thread safe and not a feature
to base product on.

---------

Signed-off-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>
Co-authored-by: Sergey Semenov <sergey.semenov@intel.com>