[SYCL] Refactor sycl::vec's operators implementation
* Don't use `sycl::vec::vector_t`, as it is planned to be removed from
the SYCL 2020 (KhronosGroup/SYCL-Docs#676). Note
that this implementation is NOT required to use it, so this PR can be
merged before the specification change.
* Use `ext_vector_type`-based optimized implementation whenever it's
available and not on device only.

This is a recommit of https://github.com/intel/llvm/pull/16529 with an
additional `#if __clang_major__ >= 20` guard around `static_assert`
on the expression that wasn't constant in clang-19.