[SYCL][E2E] Don't cover full combinations in `vec_binary_scalar_order*`
First, full coverage belongs to CTS, not to unit tests. Second,
implementation is heavily templated and there isn't much differences
potential for many intger types or vector sizes.

This PR eliminates test cases that are unlikely to excercise the code
not covered by the remaining ones. The purpose of this is to reduce the
time these tests takes as they're very close to be a bottleneck on
machines with lots of cores where total wall clock time is defined by
the longest tests in the suite.