[SYCL][NFC] Remove dead arg and check (#16534)
The JITCompilationIsRequired in a selection of build functions default
to false and are never set by any callers. Given this, the argument can
be removed. As a side-effect of this, the CheckJITCompilationForImage
function will return immediately if the JITCompilationIsRequired was
false, and since it always was the function is obsolete and can be
removed.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>