[SYCL][RTC] Use program manager (#16316)
The idea is to assign a per-compilation prefix (e.g., `rtc_42$`) to all
offload entries in the `sycl_device_binaries` datastructure before
feeding them into `ProgramManager::addImages`, resulting in unique
`kernel_id`s as far as the PM is concerned. When querying the PM for the
device images to construct the kernel bundle, I look for kernel IDs
starting with the current prefix, which should reliably return only the
device images corresponding to current compilation request.

Note that the actual kernel names don't change, i.e. `__sycl_kernel_foo`
keeps that name even though the PM might know it as
`rtc_42$__sycl_kernel_foo`. The prefix is stored inside the bundle, and
prepended to the requested kernel name in the
`ext_onapi_[has|get]_kernel(string)` methods.

Kernel objects are also obtained via the program manager. Compared to
creating the UR kernel from the selected device image's UR program
directly, this approach ensures eliminated arguments are handled
correctly. Hence, I was able to drop the previously mandatory
`-fno-sycl-dead-args-optimization` from the pipeline.

The compilation pipeline and extended kernel bundle now support multiple
device images. To test this, I added support for the
`-fsycl-device-code-split=` option, and apply it to one of the
compilations in the E2E test.

The persistent cache is circumvented for now for the `sycl_jit` language
(lack of suitable on-disk format), but should be brought back in the
future.

---------

Signed-off-by: Julian Oppermann <julian.oppermann@codeplay.com>
Co-authored-by: Lukas Sommer <lukas.sommer@codeplay.com>