[SYCL][E2E] Introduce `%{sycl_target_opts}` substitution
...and use it to pass `amd_arch` options to compilation only when
compiling for AMD triples instead of unconditional usage in `%clangxx`
substitution.

That would avoid passing the option unnecessary when running `llvm-lit`
with multiple target devices and a given test isn't supported for HIP.
We might not have such configuration presently, but the same logic will
be applied for "split" build/run of E2E tests and we'd definitely want
to build tests for all the possible targets at once.

Another use case for it is the "spirv backend" testing where we'd want
to pass extra option for any compilations targeting spirv triple.