[libclc] Reland 7e6a73959ae97b1f9476a90290a492ba90cb950d (#16606)
This upstream libclc commit was dropped during a pulldown. This relands
it, adapting it to our downstream libclc framework. Most of the upstream
commit was to the OpenCL builtins, which we don't directly make use of
for DPC++.

----

libclc: increase fp16 support (#98149)
Increase fp16 support to allow clspv to continue to be OpenCL compliant
following the update of the OpenCL-CTS adding more testing on math
functions and conversions with half.

Math functions are implemented by upscaling to fp32 and using the fp32
implementation. It garantees the accuracy required for half-precision
float-point by the CTS.