[CI] Use `intel_drivers:alldeps` container on `build_e2e_tests` precommit step (#16634)
`ubuntu2404_intel_drivers:latest` container was used here previously
because there was no Ubuntu 24 version of the `alldeps` container, and
we were only building for the `spir64` triple. This container has since
been added so we can use it instead. This will allow us to compile for
the other triples in the `build-only` step in the future.