[SYCL][Graph][E2E] Reduce work in intensive update tests running (#16632)
On certain setups these two tests take a long time to execute, impairing
the ability to run the whole suite of graph E2E tests in lit.

Reduce the number of graph execution iterations used in tests to
alleviate this situation. There is an issue with free-function kernels
taking longer that the non-free function equivalent, documented in
CMPLRLLVM-64841, so only use a single execution iteration for now.