Remove command-buffer command handle ref counting
The handles to individual commands returned from command-buffer
append entry-points currently need reference counted by the user.
However this isn't a model that maps to OpenCL/CUDA/HIP/Level-Zero
where the lifetime of commands is tied to the lifetime of the parent
command-buffer/graph/command-list.

Remove this idiom from the command-buffer UR API, adapters, and CTS.
The APIs that are removed are
* `urCommandBufferCommandGetInfoExp`
* `urCommandBufferReleaseCommandExp`
* `urCommandBufferRetainCommandExp`