[SYCL] Remove `IsDeprecatedDeviceCopyable` (#16615)
This was discussed in
https://github.com/intel/llvm/pull/15342#discussion_r1751868640 and the
consensus seemed to be that we should drop it right away in a separate
PR, do it here.

Technically, it is a breaking change that could also be considered a
bugfix. An example of a class failing the updated check is

```
struct Kernel {
    Kernel(int);
    Kernel(const Kernel&) = default;
    Kernel& operator=(const Kernel&) { return *this; } // non-trivial
};
```

An additional minor reason (other than not being SYCL-conformant) to
drop it right away is to save a tiny bit of compile time that is
currently used to support something violating the spec.

This required some fixes in the reductions implementation to make sure
the kernel we submit internally are actually device copyable.