[SYCL][Matrix] Propagate constexpr matrix layout even with O0 (#16628)
Per SPIR-V specification Layout of a matrix must be a constant
instruction aka a constexpr or specialization constant. Meanwhile in
SYCL headers layout is passed as a parameter to joint_matrix_load
function, so even if that layout is a constant expression in the user's
code - it's not possible to prove that to the compiler, so constant
propagation will happen only after inlining, not in AST. That means,
that with O0 layout would remain to be a runtime variable in LLVM IR.

SYCL matrix layout is being mapped on SPIR-V matrix layout by
joint_matrix_layout_to_spv function. This patch adds routine that finds
calls to this function and replaces them with the found constant. To
help this routine always_inline attribute was removed from
joint_matrix_layout_to_spv function.

---------

Signed-off-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>