Fix invalid use of dlopen()
When attempting to get a handle for the `libOpenCL.so` library in order
to call `dlsym()` invalid arguments were set so `dlopen()` always
returned null. This just happened to work due to the value of
`RTLD_DEFAULT` being 0 so the `dlsym()` calls succeeded. This patch
removes the invalid `dlopen()` usage and explicitly uses `RTLD_DEFAULT`
instead of the null `handle`.