[SYCL][E2E] Implement support for REQUIRES in build-only mode
In general we cannot separate
build only features from features that only have an effect during
execution. Consider a `REQUIRES` expression containing a mix of runtime
and build time features. If the expression evaluates to false in
build-only we would have to solve a Satisfiability (SAT) problem to
check if it only failed due to missing run-time features.
The below design is motivated by this and the fact that current
`REQUIRES`, `UNSUPPORTED` and `XFAIL` lines assume that both run-time
(per-device) and build only features are available.

Add a new directive `ENABLE_RUN_REQUIRES: true`, which, if present, changes
the evaluation of `REQUIRES`, `UNSUPPORTED` and `XFAIL` directives to
only consider device-agnostic features. This allows them to be correctly
evaluated during `test-mode="build-only"` too. To express per-device
(run-time) requirements the directives `RUN_REQUIRES`, `RUN_UNSUPPORTED`
and `RUN_XFAIL` can be added.

This commit does not yet change the documentation of the `test-mode`
parameter in README.md, if this looks like a good direction I can do
that too with help on the wording.