[UR] Replace loader handles with field at start of handle data
All handles from all backends are now required to implement `ddi_getter` and their first field must be a pointer to a
`ur_ddi_table_t` (which also implies that they must not have a vtable).

Instead of wrapping handles in a special wrapper type, we instead query the DDI table stored in the handle itself. This
simplifies the loader greatly.