[CI] Remove `sycl-devops-pr/**` hack
It was originally added in https://github.com/intel/llvm/pull/10002,
back when we used `pull_request_target` for pre-commit trigger. Now that
`pull_request` is used it isn't (as) necessary.

I'm also adding a `workflow_dispatch` trigger for the post-commit job
for cases when somebody wants to test extensive non-CI changes with
post-commit before the merge. I think that will work but it's impossible
to verify before merging this due to how GHA work.

This will address the following usage scenario:
  * Push CI-related changes to `sycl-devops-pr` (possibly several
    iterations)
  * Result are ok, open a PR from that branch
    - extra post-commit task is spawned for the commit that has been
      tested already (this duplication will remain, unfortunately)
  * Address review feedback with a new upload
    - two post-commit jobs are automatically launched (after this PR, no
      manual dispatch would happen here, leaving only automatic
      pre-commit testing of the post-commit workflow)