[CI] Fix race condition by unsetting `merge_ref` for Windows tasks
That input parameter was introduced to `sycl-windows-build.yml` in
https://github.com/intel/llvm/pull/16560 but it's unclear why.

Note that linux version has had it for quite some time, but we started
unsetting it explicitly in https://github.com/intel/llvm/pull/10720. The
reason we had it for Linux in the first place is because `GITHUB_SHA`
pointed to the PR's source branch for the `pull_request_target` trigger
that we used at that time. With switching to `pull_request` trigger it
became unnecessary as `GITHUB_SHA` points to the merge commit already.

This change should fix current race condition when a post-commit CI job
can test change landed after the current commit if there is a delay with
runner allocation.