[SYCL][E2E] Add logic to react to `REQUIRED`/`UNSUPPORTED` in `build-only` (#16725)
Adds logic for evaluating `REQUIRES`/`UNSUPPORTED` statements on
`build-only` mode by "ignoring" features in these statements that do not
affect the compilation (Everything other than OS, triple, and
SDK/libraries).

More precisely, the ability to ignore features is implemented by
extending the boolean expressions to use a third boolean state -
`ignore`. If a particular sub-expression includes an `ignore`, and if
its result could be changed by setting that `ignore` to either `true` or
`false`, then the result of that sub-expression is set to `ignore`. For
example `ignore || true = true`, but `ignore || false = ignore`, this is
because in the first example there would be no way to set the ignore
such that the result is anything other than `true`, while in the second
example the result is dependent on the "actual" value of the `ignore`.

If the resulting value of a `REQUIRES` predicate is `ignore` we
interpret that as `true` (The requirements were met), on the other hand
for `UNSUPPORTED` predicates we would interpret `ignore` as `false`
instead (The unsupported criteria was not met).

The triples that can be used for a given test are then selected by
evaluating the `REQUIRES`/`UNSUPPORTED` with the set of features + the
target feature corresponding to each triple (mirroring the way we select
devices), while ignoring all features that do not affect compilation.

The target features available are the following: `target-spir`,
`target-amd`, `target-nvidia`, and `target-native_cpu`. Each of these
map to a triple.

Similarly to how `XFAIL` is handled when using multiple devices if we
are compiling for multiple triples, and a single triple is marked as
`XFAIL`, then it is treated as unsupported instead.

The available target triples in `build-only` mode is determined through
the use of the `sycl_build_targets` lit param (i.e., `--param
sycl_build_targets=spir;amd`). This is currently set to only `spir`, as
the changes in test markup included in this pr do not take into account
building for `nvptx` and `amdgcn` triples. In `run-only` and `full` mode
the available triples are determined via the available devices.