[SYCL][E2E] Make layout_array.cpp more robust to ordering
The layout_array test is currently prone to failure if any of the types
it checks are instantiated by the headers rather than by the test
itself. This means that unrelated changes to the SYCL headers may make
the test fail due to one of the types being instantiated earlier than
the test instantiates other types that go before it. To avoid this
issue, these changes makes the test indifferent to the order of the
instantiated layouts in the generated output.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>