[SYCL][E2E] Make layout_array.cpp use specific headers (#16822)
The layout_array test is currently prone to failure if any of the types
it checks are instantiated by the headers rather than by the test
itself. This means that unrelated changes to the SYCL headers may make
the test fail due to one of the types being instantiated earlier than
the test instantiates other types that go before it. To avoid this
issue, this commit makes the test use specific headers, reducing the
chance of instantiations.

This commit also fixes a faulty expectation in the `nd_item` test case.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>
Co-authored-by: Daniel Skrobot <daniel.skrobot@intel.com>