[SYCL][Graph] Refine E2E test comment (#16842)
The code comment in
`Update/FreeFunctionKernels/update_with_indices_ordering.cpp` mentions
that it is unknown why the execution time is slower than the equivalent
`Update/update_with_indices_ordering.cpp` test without free functions.

This has now been investigated in CMPLRLLVM-64841 and has found to be
due to the explicit arguments in the free function signature avoiding
constant folding from lambdas. Note that we want to keep these
parameters as the test is stressing the case where multiple args are
passed to the free function.