[SYCL][E2E] Add selected targets to available features set (#16860)
This resolves two issues: 
- Currently in `build-only` XFAILs for target features don't trigger
because the target feature is not in the available features set.
- Being able to use the target feature instead of `any-target-is` for
`%if` conditions in run lines allows us to only add the bit enclosed in
the condition if that target is selected. So for example if a particular
target is available, but it is marked as xfail/unsupported for a given
test we dont add the bits enclosed in `%if target-`. On the other hand
if this were enclosed with `%if any-target-is-`, we would add it, which
is not always desirable.