[SYCL][E2E] Remove `REQUIRES: build-and-run-mode` in bindless_images tests (#16789)
As of #16725 tests can properly react to features that affect
compilation on build-only mode (i.e., libraries or OS), additionally we
can also mark if a test should only be built for a specific triple using
the `target-*` features.

This pr removes the `REQUIRES: build-and-run-mode` directives from
bindless_images tests. These tests will now be marked as unsupported in
environments that cannot build the tests due to missing the Vulkan
library, or not being on Windows. For two tests removed the `REQUIRES:
build-and-run-mode` directive by changing `REQUIRES: cuda` to `REQUIRES:
target-nvidia`, this way the tests will only compile for nvidia triple.