[SYCL][Docs] Add get_native_context call to interop_handle example
The interop_handle example in the L0 backend specification uses a L0
context handle, but it may be unclear to users how that context handle
is accessible. Since it needs to be at least associated with the
corresponding device, it seems proper to be explicit in the example
about how to get it from the interop_handle.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>