[SYCL][NFCI] Restore commit ccd73ee
Commit

> ccd73eeab34b31c7c38e9aca05ca4192fb0913b0
> [LinkerWrapper] Remove in-house handling of LTO (llvm/llvm-project#113715)

Removed a lot of dead code from `clang-linker-wrapper`. However, it has
been preserved in our fork by:

> d7c2a0fe4dea71e469fc651a87e549f4fd363202
> Merge from 'main' to 'sycl-web' (69 commits)

Due to the way we have handled a divergence from the upstream it isn't
really clear if something is our code or not, so it wasn't that hard to
make a conflict resolution mistake.

This commit removes all that dead code to make it closer to the
upstream.