[SYCL][NFCI] Unify our customizations with the community
We seem to have put all the community code into a separate branch which
made it hard to maintain (i.e. keep in sync with the upstream).

This commit outlines some common parts of downstream and upstream code
to simplify feature maintanence (by making the code closer to the
upstream overall).