`alloca`s should use the AllocaAS (#3004)
Currently, when reverse translating, we hardcode either 0 or
`SPIRAS_Private` as the address space for `alloca`s. This is not ideal
as it creates tight coupling with the current AS mappings. Furthermore,
it is not necessary, as the DataLayout holds the AllocaAS, which is what
we need. Hence, this patch changes all hardcoded callsites where we
create an `AllocaInst` to use the DataLayout specified AS. It's NFC as
it does not modify existing behaviour.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/c1a7e51fb0ea6ef