[SYCL][E2E] Change method for setting a `RUN:` line to execute at run stage (#16907)
Previously to explicitly mark that a `RUN:` line should run on the
run-stage the `%if run-mode %{...%}` markup was used. However, this
markup is misleading given that the entire line is skipped in
`build-only` mode, not just the bit enclosed in the braces. This PR
replaces this markup with `%{run-aux}`, to avoid this issue.