[SYCL][E2E] Replace `any-device-is-` with `target-` features in `%if` markup (#16889)
Using `target-` features is preferable for `%if` statements because this
feature will only evaluate to true if a device with the corresponding
build target is selected, or if we are in `build-only` and the
corresponding triple was selected to build for the test. `any-device-is`
on the other hand evaluates to true any time we have the device
available on the system, even if it is marked as XFAIL/UNSUPPORTED, and
on `build-only` it is never true because we have no devices in that
mode.