[NFCI][SYCL] Split vec's unary ops into individual mixins (#16946)
We don't want constraints on non-template hidden friend operators to be
implemented by templatizing it. It would be trivial with C++20 concepts,
but we're limited to C++17. The proposed class hierarchy for
`vecto_arith` helper had intended that, but implementation was only
partial.

We want to fix that while implementing the proposed specification
changes to the vec/swizzle so setup infrastructure to do so now, to
minimize the amount of "preview-breaking-changes" customizations later.

The idea is to split each operator into an individual unrestrained
mixin, so that the constraints could be implemented on top of it via
`detail::ApplyIf` helper. This particular PR only implements such a
change for unary operators. The rest will be implemented in a subsequent
change to ease review process.