[CI] Change AMD CPU job in Nightly to use proper label
Back when this task was added we only had runners with both AMD CPU and
GPUs. Recently, we excluded the only remaining such runner as it's GPU
isn't officially supported (we rely on the runner provided by Codeplay).
However, we can keep using this old runner for the AMD CPU testing in
Nightly as it was originally designed.