[Driver][SYCL] Do not emit mismatch warning with -fsycl-force-target
The use of -fsycl-force-target=arg allows for a user to override the
default target triple that is used to unbundle device objects from the
fat objects.  The diagnostic warning a user that the expected target
values within the given objects is not found is not valid when
-fsycl-force-target is used with the matching target value.

Avoid emitting this warning when the -fsycl-force-target=arg matches
what is found in the incoming objects.