[SYCL] Fix min/max known_identity<> wrong when -ffast-math is used. (#17028)
Technically a workaround, this fixes the problem with
`known_identity<minOp/maxOp>` being incorrect when using `-ffast-math`.
It works because we have a `__FAST_MATH__` macro that is defined when
that flag is used. Better, would be to detect `-fno-honor-infinities`.
Opening a JIRA/discussion for that, but this is a big improvement in the
interim.

See discussion https://github.com/intel/llvm/issues/13813