[SYCL][RTC] Define custom destructor for `kernel_bundle_impl` (#16702)
Adds a custom destructor to `kernel_bundle_impl` to properly clean-up
runtime information and device binaries for bundles that are
runtime-compiled from source:

- **Removal of kernels from program manager**
The new `ProgramManager::removeImages` method takes `addImages` as a
blueprint and reverses the effects of registering the given device
binaries. Currently, only a subset of the data structures in the program
manager is handled, and `assert`s are in place for the remaining
members. Device globals will be addressed after #16565 lands. AFAICT I
could clean-up most of the other members mechanically as well, but
decided against that because I can't test these features (such as
virtual functions) right now due to lack of SYCL-RTC support.

- **Free device binaries in JIT library**
Store the UR-specific data structure that backs a device binary in a map
of unique ptrs instead of a vector to make it possible to free it again
without invalidating other compilation results. Also introduce a mutex
to protect concurrent access to this new map.

- **Free raw SPIR-V binaries in JIT context**
Again replace a vector with a map of unique ptr to make it possible to
free `KernelBinary` objects in the `JITContext`. `KernelBinary` owns the
actual SPIR-V binaries; all other data structures mentioned earlier only
store the pointer return by `KernelBinary::address()`. A new function
`destroyBinary` is added to the `sycl-jit` interface.

My understanding is that `~device_image_impl` and `~kernel_impl` already
take care of releasing the underlying UR resources, hence no changes
required there.

---------

Signed-off-by: Julian Oppermann <julian.oppermann@codeplay.com>