[UR] Make command-buffer creation descriptor mandatory
As discussed in https://github.com/oneapi-src/unified-runtime/pull/2670#discussion_r1946398932 the `pCommandBufferDesc` parameter to `urCommandBufferCreateExp` is optional. However, the UR spec doesn't state what the configuration of the created command-buffer is when this isn't passed, and being optional is also inconsistent with the description parameters to urSamplerCreate & urMemImageCreate which are not optional. This PR updates the descriptor parameter to command-buffer creation to be mandatory to address these concerns.

Closes https://github.com/oneapi-src/unified-runtime/issues/2673