[libclc] Resolve bad merge; fix libspirv LIT failures
This is an alternative fix to #17052.

The merge resolution in 98f847a accidentally stopped the compilation of
all libspirv builtins, resulting in an empty library.

The libspirv directory structure is almost identical to the OpenCL
directory structure, except that it also has a native-cpu directory.
Thus we introduce a new variable to manage libspirv directories, which
starts out inheriting from the OpenCL directories.