[UR] Make each profiling info variant for urEventGetProfilingInfo optional and improve its conformance test (#17067)
Migrated from https://github.com/oneapi-src/unified-runtime/pull/2533

This patch turns all of the values returned by urEventGetProfilingInfo
to be optional and updates adapters to handle this by returning the
appropriate enum when it is not supported.

The tests have also been updated, to ensure that returning a counter of
"0" or values equal to the previous profiling event is no longer
considered a failure.