[SYCL][E2E] enable testing L0 v2 adapter
The idea is to allow passing 'level_zero_v2:gpu' as a valid
sycl_device. .py scripts are modified so that when 'v2' is detected
in the device name, appropriate env variable (to enable UR loader
to load the adapter) is set. config.available_features are also
extended with 'adapter-v2' feature which allows tests to be marked
as unsupported for a given adapter.

Since V2 adapter is not a valid backend for ONEAPI_DEVICE_SELECTOR,
When setting ONEAPI_DEVICE_SELECTOR or checking for other features,
a parsed device name is always used (which does not contain 'v2' term).