Remove unnecessary and confusing unique_ptr usage
`std::unique_ptr` was being used as a scoped deleter for a UMF value,
but this was both confusing and more expensive than it needed to be.

Here we introduce the simple `OnScopeExit` class which calls a callable
when it goes out of scope, and use it here to simplify the meaning and
overhead of the CUDA adapter's UMF pool allocation.