[SYCL] Fix type traits used for shuffle operations (#17055)
* Pointer type was incorrectly treated as arithmetic type, so the case
when user passed, for example, `int*` variable to
`sycl::select_from_group` ended up mapping to non-existent built-in
`SubgroupShuffleINTEL(i32*, ...)`. This is fixed in this PR and now it
is mapped to `SubgroupShuffleINTEL(i64, ...)`

* Rename type trait
`VecTypeIsProhibitedForShuffleEmulation`->`VecTypeIsProhibitedForNativeShuffle`.
Old name was confusing. Basically trait is used to determine vec types
for which we cannot use native built-ins.

* `enable_if`s for shuffle overloads were not completely mutually
exclusive leading to compilation error with ambiguous definitions for
some types. For example, `marray<bfloat16, 4>` used to satisfy both
`NonScalar` and `Bitcast` traits. So, refactoring was done to make them
mutually exclusive and more clear.